Atom: a lot of commented debugprints, hack to fix race conditions between socketthread and asyncsocket, counter for socketthreads id instead of time Display.js: timout for going to main menu, minor changes
git-svn-id: https://svn.arune.se/svn/HomeAutomation@1203 e69ec2e4-db17-0410-bd4a-d6f28f89cd0d

+++ trunk/PcSoftware/Atom/src/VM/virtualmachine.h
@@ -115,7 +115,7 @@ private:
 	ThreadSafeQueue<Expression> myExpressions;
 
 	map<int, IntervalThread*> myIntervalThreads;
-	map<int, SocketThread*> mySocketThreads;
+	map<unsigned long int, SocketThread*> mySocketThreads;
 
 	CommandThread myCommandThread;
 };
